app/Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3
python-multipart==0.0.6
groq==0.4.0 